@@ -9,14 +9,15 @@
         "@fortawesome/free-solid-svg-icons": "^5.15.4",
         "@fortawesome/react-fontawesome": "^0.1.16",
         "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
+        "@react-pdf/renderer": "^4.2.2",
         "@svgr/webpack": "^6.5.1",
         "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^12.0.0",
+        "@testing-library/react": "^16.2.0",
         "@testing-library/user-event": "^13.2.1",
         "@types/jest": "^27.0.3",
         "@types/node": "^17.0.4",
-        "@types/react": "^17.0.38",
-        "@types/react-dom": "^17.0.11",
+        "@types/react": "^19.0.10",
+        "@types/react-dom": "^19.0.4",
         "@types/react-router-dom": "^5.3.2",
         "babel-jest": "^27.4.2",
         "babel-loader": "^8.2.3",
@@ -50,10 +51,11 @@
         "postcss-preset-env": "^7.0.1",
         "prettier": "^2.5.1",
         "prompts": "^2.4.2",
-        "react": "^17.0.2",
+        "react": "^19.0.0",
         "react-app-polyfill": "^3.0.0",
         "react-dev-utils": "^12.0.0",
-        "react-dom": "^17.0.2",
+        "react-dom": "^19.0.0",
+        "react-pdf": "^9.2.1",
         "react-refresh": "^0.11.0",
         "react-router-dom": "^6.2.1",
         "react-select": "^5.8.2",
